requirements.txt
avro==1.10.2
pulsar_client==3.4.0
faster_whisper==0.10.0
requests==2.31.0